influxdb/pyconsumer/requirements.txt
numpy==1.15.1
opencv_python==3.4.3.18
confluent_kafka==0.11.6
python-dotenv==0.10.1
